api/tutorx/requirements.txt
asgiref==3.8.1                                                               
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==21.2.0
h11==0.14.0
packaging==24.0
pillow==10.3.0
sqlparse==0.4.4
tzdata==2024.1
uvicorn==0.29.0